Refer description
- substantially reworked `render_polar_plot` method, now uses a
`bullseye` argument that determines the size of the bullseye on the plot
- addresses #16
- now supports a `polar_rings` config option to set the number of rings
to be displayed on the polar plot. Default is 5. The number of rings
reaches into many parts of the code not just the rendering of the polar
grid, there are some places where it is still hard coded. Not fully
tested as yet.
- refactored the wind rose `render_plot` method given the ability to
scale the bullseye, think it is a lot clearer and simpler now